LOG-3949: increase max_read_bytes for container logs to release file handles sooner
