+++ internal/generator/vector/source/kubernetes_logs.go
@@ -19,6 +19,7 @@ func (kl KubernetesLogs) Template() string {
 # {{.Desc}}
 [sources.{{.ComponentID}}]
 type = "kubernetes_logs"
+max_read_bytes = 3145728
 glob_minimum_cooldown_ms = 15000
 auto_partial_merge = true
 exclude_paths_glob_patterns = {{.ExcludePaths}}
